docker: tar: Accept digested destination reference
Allow specifying of a digested reference as a destination reference in
the case of copying image to `docker-daemon`.
The digested reference is treated as a tag in this case because no
content hash match happens. The top level index manifest is not present
in the `oci` store so it's not possible to perform it at all.